% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micrometViz.R
\name{microMetViz}
\alias{microMetViz}
\title{R shiny app for visualizing daat contained within MSHMIcroMet package}
\usage{
microMetViz()
}
\description{
R shiny app for visualizing the location and metadata for HOBO deployments on Mount St. Helens and time series of temperature and relative humidity measurements collected during each deployment
}
\examples{
# microMetViz()

}
